Rebased and pulled new fixer into fixer file
